fix: satisfy trace overlay lock audit
