+++ src/core/extension/trace/mod.rs
@@ -9,6 +9,7 @@
 //! milestones.
 
 pub mod baseline;
+mod overlay_lock;
 pub mod parsing;
 pub mod report;
 pub mod run;
@@ -17,6 +18,10 @@ pub mod spans;
 use crate::component::Component;
 use crate::extension::{ExtensionCapability, ExtensionExecutionContext};
 
+pub use overlay_lock::{cleanup_stale_trace_overlay_locks, list_trace_overlay_locks};
+pub use overlay_lock::{
+    TraceOverlayLockCleanupResult, TraceOverlayLockRecord, TraceOverlayLockStatus,
+};
 pub use parsing::{parse_trace_list_str, parse_trace_results_file};
 pub use parsing::{
     TraceArtifact, TraceAssertion, TraceEvent, TraceList, TraceScenario, TraceStatus,
@@ -28,14 +33,8 @@ pub use report::{
     TraceCompareSpanOutput, TraceOverlayLocksOutput, TraceRunOrderEntryOutput,
 };
 pub use report::{push_overlay_markdown, render_markdown};
-pub use run::{
-    cleanup_stale_trace_overlay_locks, list_trace_overlay_locks, run_trace_list_workflow,
-    run_trace_workflow, TraceListWorkflowArgs,
-};
-pub use run::{
-    TraceOverlayLockCleanupResult, TraceOverlayLockRecord, TraceOverlayLockStatus,
-    TraceRunWorkflowArgs, TraceRunWorkflowResult,
-};
+pub use run::{run_trace_list_workflow, run_trace_workflow, TraceListWorkflowArgs};
+pub use run::{TraceRunWorkflowArgs, TraceRunWorkflowResult};
 
 pub fn resolve_trace_command(
     component: &Component,
